refactor(cli): Now supports mock and server
